fixed
fixed it by changing where the page "looked" after the price.

Reason : The html file wanted to lock at the secound part of the database element but when it loocked there, there was no price there where description so it didnt show enything instead. But no the price is at the placholder and where the price is supoed to be.